backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.9
alembic==1.15.2
pydantic==2.11.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.5.0
redis==5.2.1
pytest==8.3.5
python-dotenv==1.0.1
pgpy==0.6.0
